Updating product

+++ springmvc/src/main/java/guru/springframework/domain/Product.java
@@ -1,18 +1,16 @@
 package guru.springframework.domain;
 
-import java.math.BigDecimal;
-
 public class Product {
-    private int id;
+    private Integer id;
     private String description;
-    private BigDecimal price;
+    private double price;
     private String imageURL;
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -24,11 +22,11 @@ public class Product {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
